requirements.txt
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.82.0
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
deprecation==2.1.0
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Cors==4.0.0
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
postgrest==0.16.3
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.4
six==1.16.0
sniffio==1.3.1
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
typing_extensions==4.11.0
websockets==12.0
Werkzeug==3.0.2
zipp==3.18.1
Flask-BasicAuth==0.2.0
datetime